modificaciones y prueba de endpoint

+++ src/main/java/com/SySTomateAlgo/TomateAlgo/Controllers/ProductController.java
@@ -1,12 +1,15 @@
 package com.SySTomateAlgo.TomateAlgo.Controllers;
 
+import com.SySTomateAlgo.TomateAlgo.Entities.Client;
 import com.SySTomateAlgo.TomateAlgo.Entities.Product;
 import com.SySTomateAlgo.TomateAlgo.Services.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/products")
@@ -21,7 +24,11 @@ public class ProductController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Product> findById(@PathVariable Long id){
+    public ResponseEntity<?> findById(@PathVariable Long id){
+        Optional<Product> exitProduct =productService.findById(id);
+        if (exitProduct.isEmpty()){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Producto no encontrado");
+        }
         return productService.findById(id)
                 .map(ResponseEntity::ok)
                 .orElse(ResponseEntity.notFound().build());
@@ -34,11 +41,16 @@ public class ProductController {
 
     @PutMapping("/{id}")
     public ResponseEntity<Product> update(@PathVariable Long id, @RequestBody Product product){
+
         return ResponseEntity.ok(productService.update(id, product));
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<String> delete(@PathVariable Long id){
+        Optional<Product> product =productService.findById(id);
+        if (product.isEmpty()){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Producto no encontrado");
+        }
         productService.delete(id);
         return ResponseEntity.ok("Producto Eliminado correctamente");
     }
